Clean up overloaded operators
Instead of assuming a bool result for each of the comparison operators,
we'll use auto.  I can't think of a good reason why they wouldn't be
auto, but this keeps the code generic in case somebody has done a weird
specialization for the underlying type.

I also changed the static_casts to instead use operator*.  This
shouldn't have any impact on the code, but it makes it a bit cleaner (I
added operator* after the operator overloading funcitons were already
written).