Reorder arguments in helper macros
If the wrapped type needs a comma (e.g., std::pair), the helper macros
require an extra type alias prior to use--this defeats the point of the
macros making it simpler.  By putting the backing type last, I can just
pass everything after the type name as the backing type, and that avoids
the problem entirely.