Experimental work to support something akin to ruby mixins
To avoid users writing boilerplate code, I'd like to naturally support
operator overloading for things like adding two Widths together.  I
can't think of a way to make it typesafe and generic at the same time,
so the idea is the end-user would specify the specific types that it's
safe to operate with and biztonsag stamps out templated operator X
member functions.

This work is *highly* experimenetal, and I wouldn't be surprised if the
current version is breaking all sorts of rules with reinterpret_cast.