Create mixins for minus, times, and divides
These are all based on Plus with the operators changed.  Everything
seems to be working, but gcc is barfing on operator*, claiming it's
ambiguous when I try to use it.

Since clang is fine and the calls don't look ambiguous to me (unary vs
binary operator*), I'm not going to worry about the changes for now,
especially since this work is all still experimental.