New approach for arithmetic mixins
The old style depended on stamping out different versions of the
operator overloads using marcros, and I was running into all sorts of
problems.  It's eitehr a stroke of genius or insanity, but I think I can
simplify things using a pattern more akin to type traits.

This commit only has plus working, but it should scale to support the
other operators.