Improve safe string example
The safe string example now has comments to explain what's going on, and
I made the code more generic to support the idea of Safe/Unsafe versions
of any possible type.  This latter change is more in line with what Joel
is trying to suggest in his article that inspired biztonsag in the first
place.